@@ -1,25 +1,25 @@
 module github.com/equinor/radix-cost-allocation-api
 
-go 1.18
+go 1.19
 
 require (
-	github.com/coreos/go-oidc/v3 v3.2.0
-	github.com/denisenkom/go-mssqldb v0.12.2
+	github.com/coreos/go-oidc/v3 v3.5.0
+	github.com/denisenkom/go-mssqldb v0.12.3
 	github.com/equinor/radix-common v1.2.9
-	github.com/go-openapi/errors v0.20.2
-	github.com/go-openapi/runtime v0.23.1
-	github.com/go-openapi/strfmt v0.21.3
-	github.com/go-openapi/swag v0.22.0
+	github.com/go-openapi/errors v0.20.3
+	github.com/go-openapi/runtime v0.25.0
+	github.com/go-openapi/strfmt v0.21.7
+	github.com/go-openapi/swag v0.22.3
 	github.com/go-openapi/validate v0.22.0
 	github.com/golang/mock v1.6.0
 	github.com/gorilla/mux v1.8.0
-	github.com/prometheus/client_golang v1.13.0
+	github.com/prometheus/client_golang v1.14.0
 	github.com/rakyll/statik v0.1.7
-	github.com/rs/cors v1.8.2
+	github.com/rs/cors v1.8.3
 	github.com/sirupsen/logrus v1.9.0
 	github.com/spf13/pflag v1.0.5
-	github.com/stretchr/testify v1.8.0
-	github.com/urfave/negroni/v2 v2.0.2
+	github.com/stretchr/testify v1.8.2
+	github.com/urfave/negroni/v3 v3.0.0
 	gorm.io/driver/sqlserver v1.3.2
 	gorm.io/gorm v1.23.8
 )
